integration: Move traceloop test to a dedicated file.
