+++ integration/inspektor-gadget/integration_test.go
@@ -238,57 +238,3 @@ func testMain(m *testing.M) int {
 func TestMain(m *testing.M) {
 	os.Exit(testMain(m))
 }
-
-func TestTraceloop(t *testing.T) {
-	ns := GenerateTestNamespaceName("test-traceloop")
-
-	t.Parallel()
-
-	commands := []*Command{
-		CreateTestNamespaceCommand(ns),
-		{
-			Name: "StartTraceloopGadget",
-			Cmd:  "$KUBECTL_GADGET traceloop start",
-		},
-		{
-			Name: "WaitForTraceloopStarted",
-			Cmd:  "sleep 15",
-		},
-		{
-			Name: "RunTraceloopTestPod",
-			Cmd:  fmt.Sprintf("kubectl run --restart=Never -n %s --image=busybox multiplication -- sh -c 'RANDOM=output ; echo \"3*7*2\" | bc > /tmp/file-$RANDOM ; sleep infinity'", ns),
-		},
-		{
-			Name: "WaitForTraceloopTestPod",
-			Cmd:  fmt.Sprintf("sleep 5 ; kubectl wait -n %s --for=condition=ready pod/multiplication ; kubectl get pod -n %s ; sleep 2", ns, ns),
-		},
-		{
-			Name:           "CheckTraceloopList",
-			Cmd:            "sleep 20; $KUBECTL_GADGET traceloop list | grep multiplication",
-			ExpectedRegexp: "multiplication",
-		},
-		{
-			Name:           "CheckTraceloopShow",
-			Cmd:            "CONTAINER_ID=$($KUBECTL_GADGET traceloop list | grep multiplication | awk '{ print $5 }'); $KUBECTL_GADGET traceloop show $CONTAINER_ID | grep -C 5 write",
-			ExpectedRegexp: `bc\s+write\s+fd=\d+,\s+buf="42`,
-		},
-		{
-			Name:    "PrintTraceloopList",
-			Cmd:     "$KUBECTL_GADGET traceloop list",
-			Cleanup: true,
-		},
-		{
-			Name:    "StopTraceloopGadget",
-			Cmd:     "$KUBECTL_GADGET traceloop stop",
-			Cleanup: true,
-		},
-		{
-			Name:    "WaitForTraceloopStopped",
-			Cmd:     "sleep 15",
-			Cleanup: true,
-		},
-		DeleteTestNamespaceCommand(ns),
-	}
-
-	RunTestSteps(commands, t)
-}
